+++ screeps-profiler.js
@@ -1,9 +1,12 @@
 'use strict';
 
+const ROOT_NAME = '(root)';
+const TICK_NAME = '(tick)';
+
 let usedOnStart = 0;
 let enabled = false;
 let depth = 0;
-let parentFn = '(tick)';
+let parentFn = TICK_NAME;
 
 class ProfilerError extends Error {}
 
@@ -16,8 +19,9 @@ try {
 }
 
 function setupProfiler() {
-  depth = 0; // reset depth, this needs to be done each tick.
-  parentFn = '(tick)';
+  depth = 0; // reset variables, this needs to be done each tick.
+  parentFn = TICK_NAME;
+
   Game.profiler = {
     stream(duration, filter) {
       setupMemory('stream', duration || 10, filter);
@@ -56,23 +60,24 @@ function setupProfiler() {
 }
 
 function setupMemory(profileType, duration, filter) {
-  resetMemory();
   const disableTick = Number.isInteger(duration) ? Game.time + duration : false;
-  if (!Memory.profiler) {
-    Memory.profiler = {
-      map: {},
-      totalTime: 0,
-      enabledTick: Game.time + 1,
-      disableTick,
-      type: profileType,
-      filter,
-    };
-  }
+
+  Memory.profiler = {
+    map: {},
+    totalTime: 0,
+    totalOKs: 0,
+    totalNOKs: 0,
+    enabledTick: Game.time + 1,
+    disableTick,
+    type: profileType,
+    filter,
+  };
+
   console.log(`Profiling type ${profileType} started at ${Game.time + 1} for ${duration} ticks`);
 }
 
 function resetMemory() {
-  Memory.profiler = null;
+  Memory.profiler = undefined;
 }
 
 function overloadCPUCalc() {
@@ -108,12 +113,16 @@ function wrapFunction(name, originalFunction) {
     const profiler = wrappedFunction.__profiler;
     if (profiler.isProfiling()) {
       const nameMatchesFilter = name === getFilter();
-      const start = Game.cpu.getUsed();
       if (nameMatchesFilter) {
         depth++;
       }
       const curParent = parentFn;
       parentFn = name;
+
+      const startOKs = Memory.profiler.totalOKs;
+      const startNOKs = Memory.profiler.totalNOKs;
+      const startT = Game.cpu.getUsed();
+
       let result;
       if (this && this.constructor === wrappedFunction) {
         // eslint-disable-next-line new-cap
@@ -121,14 +130,28 @@ function wrapFunction(name, originalFunction) {
       } else {
         result = originalFunction.apply(this, arguments);
       }
-      parentFn = curParent;
-      if (depth > 0 || !getFilter()) {
-        const end = Game.cpu.getUsed();
-        profiler.record(name, end - start, parentFn);
+
+      const endT = Game.cpu.getUsed();
+
+      if (Profiler.intents.has(name)) {
+        const isOK = result === 0;
+        Memory.profiler.totalOKs += isOK ? 1 : 0;
+        Memory.profiler.totalNOKs += isOK ? 0 : 1;
       }
+
+      const endOKs = Memory.profiler.totalOKs;
+      const endNOKs = Memory.profiler.totalNOKs;
+
+      parentFn = curParent;
+
+      if (depth > 0 || !getFilter()) {
+        Profiler.record(name, endT - startT, endOKs - startOKs, endNOKs - startNOKs, parentFn);
+      }
+
       if (nameMatchesFilter) {
         depth--;
       }
+
       return result;
     }
 
@@ -136,6 +159,7 @@ function wrapFunction(name, originalFunction) {
       // eslint-disable-next-line new-cap
       return new originalFunction(...arguments);
     }
+
     return originalFunction.apply(this, arguments);
   }
 
@@ -216,8 +240,6 @@ function profileObjectFunctions(object, label) {
     const originalFunction = objectToWrap[functionName];
     objectToWrap[functionName] = profileFunction(originalFunction, extendedLabel);
   });
-
-  return objectToWrap;
 }
 
 function profileFunction(fn, functionName) {
@@ -277,34 +299,84 @@ const Profiler = {
 
   callgrind() {
     if (!Memory.profiler || !Memory.profiler.enabledTick) return null;
+    const POS = 1; // very fake, but improves readability
+
+    const SCALE = 1000000;
+    const INTENT_COST_SCALED = 0.2 * SCALE;
+
     const elapsedTicks = Game.time - Memory.profiler.enabledTick + 1;
-    Profiler.checkMapItem('(tick)');
-    Memory.profiler.map['(tick)'].calls = elapsedTicks;
-    Memory.profiler.map['(tick)'].time = Memory.profiler.totalTime;
-    Profiler.checkMapItem('(root)');
-    Memory.profiler.map['(root)'].calls = 1;
-    Memory.profiler.map['(root)'].time = Memory.profiler.totalTime;
-    Profiler.checkMapItem('(tick)', Memory.profiler.map['(root)'].subs);
-    Memory.profiler.map['(root)'].subs['(tick)'].calls = elapsedTicks;
-    Memory.profiler.map['(root)'].subs['(tick)'].time = Memory.profiler.totalTime;
-    let body = `events: ns\nsummary: ${Math.round(
-      Memory.profiler.totalTime * 1000000
-      )}\n`;
+
+    // fill actual call
+    Profiler.checkMapItem(TICK_NAME);
+    Memory.profiler.map[TICK_NAME].calls = elapsedTicks;
+    Memory.profiler.map[TICK_NAME].time = Memory.profiler.totalTime;
+    Memory.profiler.map[TICK_NAME].OKs = Memory.profiler.totalOKs;
+    Memory.profiler.map[TICK_NAME].NOKs = Memory.profiler.totalNOKs;
+
+    // fill "holder" of the call tree
+    Profiler.checkMapItem(ROOT_NAME);
+    Memory.profiler.map[ROOT_NAME].calls = 1;
+    Memory.profiler.map[ROOT_NAME].time = Memory.profiler.totalTime;
+    Memory.profiler.map[ROOT_NAME].OKs = Memory.profiler.totalOKs;
+    Memory.profiler.map[ROOT_NAME].NOKs = Memory.profiler.totalNOKs;
+
+    // "holder" has all the costs as well, but will be subtracted in the loop
+    Profiler.checkMapItem(TICK_NAME, Memory.profiler.map[ROOT_NAME].subs);
+    Memory.profiler.map[ROOT_NAME].subs[TICK_NAME].calls = elapsedTicks;
+    Memory.profiler.map[ROOT_NAME].subs[TICK_NAME].time = Memory.profiler.totalTime;
+    Memory.profiler.map[ROOT_NAME].subs[TICK_NAME].OKs = Memory.profiler.totalOKs;
+    Memory.profiler.map[ROOT_NAME].subs[TICK_NAME].NOKs = Memory.profiler.totalNOKs;
+
+    let body = '';
     for (const fnName of Object.keys(Memory.profiler.map)) {
+      // exclusive costs
       const fn = Memory.profiler.map[fnName];
+      // wall time
+      let uCPUWallOuter = fn.time * SCALE;
+      // cost for [A]ction call that returns OK
+      let uCPUIntentOuter = fn.OKs * INTENT_COST_SCALED;
+      // number of [A]ction calls that returns NOK
+      let NOKsOuter = fn.NOKs;
+
       let callsBody = '';
-      let callsTime = 0;
       for (const callName of Object.keys(fn.subs)) {
+        // costs added to caller for inclusive costs
         const call = fn.subs[callName];
-        const ns = Math.round(call.time * 1000000);
-        callsBody += `cfn=${callName}\ncalls=${call.calls} 1\n1 ${ns}\n`;
-        callsTime += call.time;
+        // wall time
+        const uCPUWallInner = call.time * SCALE;
+        uCPUWallOuter -= uCPUWallInner;
+        // cost for intent call that returns OK
+        const uCPUIntentInner = call.OKs * INTENT_COST_SCALED;
+        uCPUIntentOuter -= uCPUIntentInner;
+        // number of intent calls that returns NOK
+        const NOKsInner = call.NOKs;
+        NOKsOuter -= NOKsInner;
+
+        callsBody += `cfn=${callName}\ncalls=${call.calls} ${POS}\n${POS} `
+          + `${Math.round(uCPUWallInner)} ${Math.round(uCPUIntentInner)} ${NOKsInner}\n`;
       }
-      body += `\nfn=${fnName}\n1 ${Math.round(
-        (fn.time - callsTime) * 1000000
-        )}\n${callsBody}`;
+
+      body += `\nfn=${fnName}\n${POS} ${Math.round(uCPUWallOuter)} `
+        + `${Math.round(uCPUIntentOuter)} ${NOKsOuter}\n${callsBody}`;
     }
-    return body;
+
+    const uCPUWallTotal = Memory.profiler.totalTime * SCALE;
+    const uCPUIntentTotal = Memory.profiler.totalOKs * INTENT_COST_SCALED;
+    const NOKsTotal = Memory.profiler.totalNOKs;
+
+    const header = [];
+    header.push('# callgrind format');
+    // it seems bug in q(k)cachegrind forces that event names start with different letters
+    header.push('event: wall_uCPU : uCPU total');
+    header.push('event: intent_uCPU : uCPU [I]intent cost');
+    header.push('event: delta_uCPU = wall_uCPU - intent_uCPU: uCPU without [I]intent cost');
+    header.push('event: NOKs : [I]intents that returned !== OK');
+    header.push('events: wall_uCPU intent_uCPU NOKs');
+
+    const summary = `summary: ${Math.round(uCPUWallTotal)} `
+      + `${Math.round(uCPUIntentTotal)} ${NOKsTotal}\n`;
+
+    return header.join('\n') + summary + body;
   },
 
   output(passedOutputLengthLimit) {
@@ -412,26 +484,148 @@ const Profiler = {
     { name: 'Tombstone', val: Tombstone },
   ],
 
+  intents: new Set([
+    'Game.notify',
+    'Market.cancelOrder',
+    'Market.changeOrderPrice',
+    'Market.createOrder',
+    'Market.deal',
+    'Market.extendOrder',
+    'ConstructionSite.remove',
+    'Creep.attack',
+    'Creep.attackController',
+    'Creep.build',
+    'Creep.claimController',
+    'Creep.dismantle',
+    'Creep.drop',
+    'Creep.generateSafeMode',
+    'Creep.harvest',
+    'Creep.heal',
+    'Creep.move',
+    'Creep.notifyWhenAttacked',
+    'Creep.pickup',
+    'Creep.rangedAttack',
+    'Creep.rangedHeal',
+    'Creep.rangedMassAttack',
+    'Creep.repair',
+    'Creep.reserveController',
+    'Creep.signController',
+    'Creep.suicide',
+    'Creep.transfer',
+    'Creep.upgradeController',
+    'Creep.withdraw',
+    'Flag.remove',
+    'Flag.setColor',
+    'Flag.setPosition',
+    'OwnedStructure.destroy',
+    'OwnedStructure.notifyWhenAttacked',
+    'PowerCreep.delete',
+    'PowerCreep.drop',
+    'PowerCreep.enableRoom',
+    'PowerCreep.move',
+    'PowerCreep.notifyWhenAttacked',
+    'PowerCreep.pickup',
+    'PowerCreep.renew',
+    'PowerCreep.spawn',
+    'PowerCreep.suicide',
+    'PowerCreep.transfer',
+    'PowerCreep.upgrade',
+    'PowerCreep.usePower',
+    'PowerCreep.withdraw',
+    'Room.createConstructionSite',
+    'Room.createFlag',
+    'RoomPosition.createConstructionSite',
+    'RoomPosition.createFlag',
+    'Structure.destroy',
+    'Structure.notifyWhenAttacked',
+    'StructureController.activateSafeMode',
+    'StructureController.unclaim',
+    'StructureExtension.destroy',
+    'StructureExtension.notifyWhenAttacked',
+    'StructureExtractor.destroy',
+    'StructureExtractor.notifyWhenAttacked',
+    'StructureFactory.destroy',
+    'StructureFactory.notifyWhenAttacked',
+    'StructureFactory.produce',
+    'StructureInvaderCore.destroy',
+    'StructureInvaderCore.notifyWhenAttacked',
+    'StructureKeeperLair.destroy',
+    'StructureKeeperLair.notifyWhenAttacked',
+    'StructureLab.destroy',
+    'StructureLab.notifyWhenAttacked',
+    'StructureLab.boostCreep',
+    'StructureLab.reverseReaction',
+    'StructureLab.runReaction',
+    'StructureLab.unboostCreep',
+    'StructureLink.destroy',
+    'StructureLink.notifyWhenAttacked',
+    'StructureLink.transferEnergy',
+    'StructureNuker.destroy',
+    'StructureNuker.notifyWhenAttacked',
+    'StructureNuker.launchNuke',
+    'StructureObserver.destroy',
+    'StructureObserver.notifyWhenAttacked',
+    'StructureObserver.observe',
+    'StructurePowerBank.destroy',
+    'StructurePowerBank.notifyWhenAttacked',
+    'StructurePowerSpawn.destroy',
+    'StructurePowerSpawn.notifyWhenAttacked',
+    'StructurePowerSpawn.processPower',
+    'StructurePortal.destroy',
+    'StructurePortal.notifyWhenAttacked',
+    'StructureRampart.destroy',
+    'StructureRampart.notifyWhenAttacked',
+    'StructureRampart.setPublic',
+    'StructureRoad.destroy',
+    'StructureRoad.notifyWhenAttacked',
+    'StructureSpawn.destroy',
+    'StructureSpawn.notifyWhenAttacked',
+    'StructureSpawn.createCreep',
+    'StructureSpawn.spawnCreep',
+    'StructureSpawn.recycleCreep',
+    'StructureSpawn.renewCreep',
+    // StructureSpawn.Spawning.cancel
+    // StructureSpawn.Spawning.setDirections
+    'StructureStorage.destroy',
+    'StructureStorage.notifyWhenAttacked',
+    'StructureTerminal.destroy',
+    'StructureTerminal.notifyWhenAttacked',
+    'StructureTerminal.send',
+    'StructureTower.destroy',
+    'StructureTower.notifyWhenAttacked',
+    'StructureTower.heal',
+    'StructureTower.attack',
+    'StructureTower.repair',
+    'StructureWall.destroy',
+    'StructureWall.notifyWhenAttacked',
+  ]),
+
   checkMapItem(functionName, map = Memory.profiler.map) {
     if (!map[functionName]) {
       // eslint-disable-next-line no-param-reassign
       map[functionName] = {
         time: 0,
         calls: 0,
+        OKs: 0,
+        NOKs: 0,
         subs: {},
       };
     }
   },
 
-  record(functionName, time, parent) {
+  record(functionName, time, OKs, NOKs, parent) {
     this.checkMapItem(functionName);
-    Memory.profiler.map[functionName].calls++;
     Memory.profiler.map[functionName].time += time;
+    Memory.profiler.map[functionName].calls++;
+    Memory.profiler.map[functionName].OKs += OKs;
+    Memory.profiler.map[functionName].NOKs += NOKs;
     if (parent) {
       this.checkMapItem(parent);
       this.checkMapItem(functionName, Memory.profiler.map[parent].subs);
-      Memory.profiler.map[parent].subs[functionName].calls++;
       Memory.profiler.map[parent].subs[functionName].time += time;
+      Memory.profiler.map[parent].subs[functionName].calls++;
+      Memory.profiler.map[parent].subs[functionName].OKs += OKs;
+      Memory.profiler.map[parent].subs[functionName].NOKs += NOKs;
     }
   },
 
@@ -515,6 +709,10 @@ module.exports = {
   enable() {
     enabled = true;
     hookUpPrototypes();
+  },
+
+  disable() {
+    enabled = false;
   },
 
   output: Profiler.output,